No big change. Modification based on code reviewer

+++ run_test.sh
@@ -5,8 +5,19 @@ export CINDER_REPO_URL="https://git.openstack.org/openstack/cinder"
 export INFORTREND_TEST_DIR="cinder/tests/unit/volume/drivers/infortrend"
 export INFORTREND_DRIVER_DIR="cinder/volume/drivers/infortrend"
 
-echo "${CINDER_DIR}"
+echo "Running Flake8..."
+flake8 $CINDER_DIR/$INFORTREND_DRIVER_DIR/
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+flake8 $CINDER_DIR/$INFORTREND_TEST_DIR/
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+echo "Complete."
+
 if [ -d "${CINDER_DIR}" ]; then
+    echo "Deleting $CINDER_DIR"
     rm -rf $CINDER_DIR
 fi
 
@@ -20,21 +31,24 @@ if [ ! -d "$CINDER_DIR/$INFORTREND_TEST_DIR" ]; then
     mkdir $CINDER_DIR/$INFORTREND_TEST_DIR
 fi
 
+echo "Copy ./infortrend/*"
 cp ./infortrend/* $CINDER_DIR/$INFORTREND_DRIVER_DIR/ -r
+echo "Copy ./test/infortrend/*"
 cp ./test/infortrend/* $CINDER_DIR/$INFORTREND_TEST_DIR/ -r
 
-if [ $(grep "infortrend" -q "$CINDER_DIR/cinder/opts.py") ]; then
+if grep "infortrend" -q "$CINDER_DIR/cinder/opts.py"; then
     echo "Driver opts already set."
 else
+    echo "Setup infortrend opts/exceptions.."
     source setupIFTdriver.sh
 fi
 
 cd $CINDER_DIR
 
-if [ $(grep "'{posargs}' --concurrency" -q "tox.ini") ]; then
+if grep "'{posargs}' --concurrency" -q "tox.ini"; then
     echo "Concurrency already set."
 else
-    # concurrency set to 4 for travisCI
+    echo "Setup concurrency=4 for travisCI.."
     sed -i "s#stestr run '{posargs}'#stestr run '{posargs}' --concurrency=4#g" tox.ini
 fi
